ResultView 에서 WinningLottos 로 분리
